Rename Card and CardDetails
Fit the approvements request
