--- src/components/cardDetails/style.css
@@ -1,13 +1,0 @@
-.sider {
-    background: #fff;
-    width: 150px;
-    text-align: right;
-
-}
-
-.siderButton{
-    margin: 4px;
-}
-.content {
-    background: #fff;
-}
