+++ src/components/cardPreview/DraggableCard.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { Draggable } from 'react-beautiful-dnd'
-import Card from './Card'
+import Card from './CardPreview'
 import './style.css'
 
 const DraggableCard = props => (
